+++ src/Cart/Events/CartDeleted.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the CartDeleted class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-03-07
+ *
+ */
+
+namespace Vanilo\Cart\Events;
+
+final class CartDeleted
+{
+}